requirements.txt
torch>=2.4.0
torchvision>=0.19.0
numpy>=2.0.0
pillow
dash
dash-bootstrap-components
plotly
gunicorn
opencv-python-headless